requirements.txt
Flask
redis
Pillow
celery
minio
